refactor: zustand selector 적용 & schedule table 컴포넌트 파일 분리 & 익명 함수 제거
