+++ package.json
@@ -25,6 +25,7 @@
     "msw": "latest",
     "react": "latest",
     "react-dom": "latest",
+    "reselect": "^5.1.1",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
