+++ src/components/ScheduleTables/DraggableSchedule.tsx
@@ -1,0 +1,86 @@
+import {
+  Box,
+  Button,
+  Popover,
+  PopoverArrow,
+  PopoverBody,
+  PopoverCloseButton,
+  PopoverContent,
+  PopoverTrigger,
+  Text,
+} from "@chakra-ui/react";
+import { CellSize, DAY_LABELS } from "../../constants";
+import { Schedule } from "../../types";
+import { useDraggable } from "@dnd-kit/core";
+import { CSS } from "@dnd-kit/utilities";
+import { ComponentProps, memo, useCallback, useMemo } from "react";
+
+export const DraggableSchedule = memo(
+  ({
+    id,
+    data,
+    bg,
+    onDeleteButtonClick,
+  }: { id: string; data: Schedule } & ComponentProps<typeof Box> & {
+      onDeleteButtonClick: (tableId: string, timeInfo: { day: string; time: number }) => void;
+    }) => {
+    const { day, range, room, lecture } = data;
+    const { isDragging, attributes, setNodeRef, listeners, transform } = useDraggable({
+      id,
+      data,
+    });
+
+    const style = useMemo(() => {
+      const leftIndex = DAY_LABELS.indexOf(day as (typeof DAY_LABELS)[number]);
+      const topIndex = range[0] - 1;
+      const size = range.length;
+
+      return {
+        position: "absolute" as const,
+        left: `${120 + CellSize.WIDTH * leftIndex + 1}px`,
+        top: `${40 + (topIndex * CellSize.HEIGHT + 1)}px`,
+        width: `${CellSize.WIDTH - 1}px`,
+        height: `${CellSize.HEIGHT * size - 1}px`,
+        transform: CSS.Translate.toString(transform),
+        zIndex: isDragging ? 10 : 1,
+      };
+    }, [day, range, transform, isDragging]);
+
+    const handleDeleteButtonClick = useCallback(() => {
+      const [tableId] = id.split(":");
+      onDeleteButtonClick(tableId, { day: data.day, time: data.range[0] });
+    }, [id, data, onDeleteButtonClick]);
+
+    return (
+      <Popover>
+        <PopoverTrigger>
+          <Box
+            bg={bg}
+            p={1}
+            boxSizing="border-box"
+            cursor="pointer"
+            ref={setNodeRef}
+            style={style}
+            {...listeners}
+            {...attributes}
+          >
+            <Text fontSize="sm" fontWeight="bold">
+              {lecture.title}
+            </Text>
+            <Text fontSize="xs">{room}</Text>
+          </Box>
+        </PopoverTrigger>
+        <PopoverContent onClick={(event) => event.stopPropagation()}>
+          <PopoverArrow />
+          <PopoverCloseButton />
+          <PopoverBody>
+            <Text>강의를 삭제하시겠습니까?</Text>
+            <Button colorScheme="red" size="xs" onClick={handleDeleteButtonClick}>
+              삭제
+            </Button>
+          </PopoverBody>
+        </PopoverContent>
+      </Popover>
+    );
+  }
+);
